feat: 프로필 편집 뷰에서 프로필 이미지 가져오기 구현
