+++ NineInTeam/NineInTeam/Assets.xcassets/View/Logout.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "rectangle.portrait.and.arrow.forward.svg",
+      "filename" : "rectangle.portrait.and.arrow.right 1.svg",
       "idiom" : "universal"
     }
   ],
